feat(db): 緯度、経度のデータ型をDoubleにする

+++ db/migrations/20240102150951_create_plan_candidate_table.sql
@@ -13,8 +13,8 @@ CREATE TABLE plan_candidate_set_meta_data
 (
     id                            CHAR(36) PRIMARY KEY,
     plan_candidate_set_id         CHAR(36)  NOT NULL,
-    latitude_start                FLOAT     NOT NULL,
-    longitude_start               FLOAT     NOT NULL,
+    latitude_start                DOUBLE    NOT NULL,
+    longitude_start               DOUBLE    NOT NULL,
     created_from_current_location BOOL      NOT NULL,
     created_at                    TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at                    TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
